[QuartzCore] Improve headers that appear generated (#40)
* [QuartzCore] Write CAFilter by hand (replace generated copy)

The header looks like it was copied from limneos headers,
where it appears to be edited (adding the comment and string externs)
by someone, despite the comment at the top. I removed the comment with
the filter types because the types change between versions. I also
removed the key-value methods because they're available from the superclass.

  html: https://developer.limneos.net/index.php?ios=3.0&framework=QuartzCore.framework&header=CAFilter.h

  text: https://developer.limneos.net/headers/3.0/QuartzCore.framework/Headers/CAFilter.h

* [QuartzCore] Move hand-written components out of QuartzCore-Structs.h

The header looks like it was copied from limneos headers,
where it appears to be edited (fixing some types, adding functions, etc.)
by someone, despite the comment at the top. I decided to keep the structs,
and import the files where the functions are moved to, for compatibility.

  html: https://developer.limneos.net/index.php?ios=3.0&framework=QuartzCore.framework&header=QuartzCore-Structs.h

  text: https://developer.limneos.net/headers/3.0/QuartzCore.framework/Headers/QuartzCore-Structs.h

* [QuartzCore] Correct typos in comments of CAFilter.h